pyproject.toml
[project]
name = "willy-project"
version = "0.1.0"
description = "Backend Django/PostGIS pour Cartronic"
authors = [
    { name = "Bikoun Ayibodo Steve William", email = "bikounayibodo@et.esiea.fr" }
]
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "Django==5.1.3",
    "djangorestframework==3.15.2",
    "whitenoise>=6.8,<7",
    "gunicorn>=23,<24",
    "uvicorn[standard]>=0.34,<1",
    "psycopg[binary]",
    "geopy==2.4.1",
    "haversine==2.9.0",
    "djangorestframework-simplejwt==5.3.1",
    "django-cors-headers==4.6.0",
    "Pillow>=11.0.0",
    "python-decouple==3.8",
    "python-dotenv==1.0.1",
    "drf-yasg==1.21.8",
    "drf-spectacular==0.27.2",
    "djangorestframework-gis==1.1",
    "django-filter==24.3",
    "pytz==2024.2",
    "celery==5.4.0",
    "redis==5.2.0",
    "channels==4.1.0",
    "channels-redis==4.2.0",
    "requests==2.32.3",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
